LayoutTests/svg/dom/viewspec-parser-2.html

<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/scripted-random.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
    var testsToRun = [
      "svgView(viewBox(1,2,3,4);preserveAspectRatio(xMaxYMin))",
      "svgView(zoomAndPan(disable))",
      "svgView(viewTarget(#myElementId))",
      ];
</script>
<script src="resources/viewspec-parser.js"></script>
</html>
