LayoutTests/svg/dom/viewspec-parser-2-expected.txt

This test fuzzes the transform parser with semi-random attribute values and dumps the results of any values that parse successfully.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Loaded: resources/viewspec-target.svg#
Parsed: undefined from: [initial view]

Loaded: resources/viewspec-target.svg#svgView(viewBox())
Parsed: undefined from: svgView(viewBox())

Loaded: resources/viewspec-target.svg#svgView(viewBox(0))
Parsed: undefined from: svgView(viewBox(0))

Loaded: resources/viewspec-target.svg#svgView(viewBox(0,0))
Parsed: undefined from: svgView(viewBox(0,0))

Loaded: resources/viewspec-target.svg#svgView(viewBox(0,0,0))
Parsed: undefined from: svgView(viewBox(0,0,0))

Loaded: resources/viewspec-target.svg#svgView(viewBox(0,0,0,0))
Parsed: undefined from: svgView(viewBox(0,0,0,0))

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatio())
Parsed: undefined from: svgView(preserveAspectRatio())

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatio(0))
Parsed: undefined from: svgView(preserveAspectRatio(0))

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatio(0,0))
Parsed: undefined from: svgView(preserveAspectRatio(0,0))

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatio(0,0,0))
Parsed: undefined from: svgView(preserveAspectRatio(0,0,0))

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatio(0,0,0,0))
Parsed: undefined from: svgView(preserveAspectRatio(0,0,0,0))

Loaded: resources/viewspec-target.svg#svgView(transform())
Parsed: undefined from: svgView(transform())

Loaded: resources/viewspec-target.svg#svgView(transform(0))
Parsed: undefined from: svgView(transform(0))

Loaded: resources/viewspec-target.svg#svgView(transform(0,0))
Parsed: undefined from: svgView(transform(0,0))

Loaded: resources/viewspec-target.svg#svgView(transform(0,0,0))
Parsed: undefined from: svgView(transform(0,0,0))

Loaded: resources/viewspec-target.svg#svgView(transform(0,0,0,0))
Parsed: undefined from: svgView(transform(0,0,0,0))

Loaded: resources/viewspec-target.svg#svgView(zoomAndPan())
Parsed: undefined from: svgView(zoomAndPan())

Loaded: resources/viewspec-target.svg#svgView(zoomAndPan(0))
Parsed: undefined from: svgView(zoomAndPan(0))

Loaded: resources/viewspec-target.svg#svgView(zoomAndPan(0,0))
Parsed: undefined from: svgView(zoomAndPan(0,0))

Loaded: resources/viewspec-target.svg#svgView(zoomAndPan(0,0,0))
Parsed: undefined from: svgView(zoomAndPan(0,0,0))

Loaded: resources/viewspec-target.svg#svgView(zoomAndPan(0,0,0,0))
Parsed: undefined from: svgView(zoomAndPan(0,0,0,0))

Loaded: resources/viewspec-target.svg#svgView(viewTarget())
Parsed: undefined from: svgView(viewTarget())

Loaded: resources/viewspec-target.svg#svgView(viewTarget(0))
Parsed: undefined from: svgView(viewTarget(0))

Loaded: resources/viewspec-target.svg#svgView(viewTarget(0,0))
Parsed: undefined from: svgView(viewTarget(0,0))

Loaded: resources/viewspec-target.svg#svgView(viewTarget(0,0,0))
Parsed: undefined from: svgView(viewTarget(0,0,0))

Loaded: resources/viewspec-target.svg#svgView(viewTarget(0,0,0,0))
Parsed: undefined from: svgView(viewTarget(0,0,0,0))

PASS successfullyParsed is true

TEST COMPLETE
